Move code from .../routes/networks to .../client for background worker
